+++ CHANGELOG.rst
@@ -4,6 +4,7 @@
 
 * record header version number in binary module and check compatibility with
   the library during startup
+* add optional output params to docs [kleis]
 
 ## Version 2.1.2 (1 March 2018)